do a better job at fixing rune logic
I'm still not wildly happy with this; there is a balance between making
a word long enough that it can be discarded as not natural language and
being able to discard something immediately that is clearly a valid token.
Discarding rune literals in const strings of binary data breaks the data
into small fragments that can then get flagged as misspellings. Given
how few characters there are in rune literal const binary data, I suspect
that entropy evaluation will not help either.